% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFeedRequirement.R
\name{calcFeedRequirement}
\alias{calcFeedRequirement}
\title{Calculate Feed Requirements}
\usage{
calcFeedRequirement()
}
\value{
MAgPIE-FEED data for Feed Requirement and corresonding weights as a
list of two MAgPIE objects
}
\description{
Provides MAgPIE-FEED data for Feed Requirement calculated in the regression
for feed (calcRegressionFEED). No changes to the content have been done.
Usually no weight needed as the data will be used in MAgPIE-FEED model which
is country based.
}
\examples{

\dontrun{ 
calcOutput("FeedRequirement")

}
}
\seealso{
\code{\link[madrat]{calcOutput}}, \code{\link{calcRegressionFEED}},
\code{\link[madrat]{readSource}}
}
\author{
Lavinia Baumstark, Isabelle Weindl
}
