% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFeeduseFoddr.R
\name{calcFeeduseFoddr}
\alias{calcFeeduseFoddr}
\title{Calculate FAO Fodder used as feed aggregated to MAgPIE_FEED sectors}
\usage{
calcFeeduseFoddr()
}
\value{
FAO fodder feed use data and corresponding weights as a list of two
MAgPIE objects
}
\description{
Provides the FAOSTAT Fodder feed use data (in mio ton) aggregated to three
MAgPIE_FEED fodder categories.
}
\examples{

\dontrun{ 
calcOutput("FeeduseFoddr")

}

}
\seealso{
\code{\link{calcProdFoddr}}, \code{\link[madrat]{calcOutput}},
\code{\link{calcFAOFodder_aggrFEED}}, \code{\link[madrat]{readSource}}
}
\author{
Isabelle Weindl
}
