% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFoodWasteRecycle.R
\name{calcFoodWasteRecycle}
\alias{calcFoodWasteRecycle}
\title{Calculate FoodWasteRecycle}
\usage{
calcFoodWasteRecycle()
}
\value{
MAgPIE object
}
\description{
Reads in information on the share of food waste recycled for feed use of the
source WirseniusPHD. Implementation of a weight calculated by
calcDemandAgriculture (y2000, food.ssp2).
}
\examples{

\dontrun{ 
calcOutput("FoodWasteRecycle")

}
}
\seealso{
\code{\link[madrat]{calcOutput}}, \code{\link{readWirseniusPHD}},
\code{\link{convertWirseniusPHD}}
}
\author{
Nele Steinmetz, Isabelle Weindl, Benjamin Leon Bodirsky
}
