% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLivstProduction.R
\name{calcLivstProduction}
\alias{calcLivstProduction}
\title{Calculate FAO Livestock Production}
\usage{
calcLivstProduction()
}
\value{
FAO Livestock Production (WM) and corresonding weights as a list of
two MAgPIE objects
}
\description{
Provides the FAO Livestock Production data aggregated to MAgPIE categories
in million ton wet matter.
}
\examples{

\dontrun{ 
calcOutput("LivstProduction")

}
}
\seealso{
\code{\link[madrat]{calcOutput}}, \code{\link[madrat]{readSource}}
}
\author{
Isabelle Weindl
}
