% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLivstSubProduction.R
\name{calcLivstSubProduction}
\alias{calcLivstSubProduction}
\title{Calculate  Production allocated to livestock production subsystems}
\usage{
calcLivstSubProduction()
}
\value{
Livestock Production (WM) and corresonding weights as a list of
two MAgPIE objects
}
\description{
Provides the FAO Livestock Production data differentiated into livestock subsystems
in million ton wet matter (based on subsystem allocation from Stefan Wirsenius).
}
\examples{

\dontrun{ 
calcOutput("LivstSubProduction")

}
}
\seealso{
\code{\link[madrat]{calcOutput}}, \code{\link{readWirseniusSubsystems}},
\code{\link[madrat]{readSource}}
}
\author{
Isabelle Weindl
}
