% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNutrientDensity.R
\name{calcNutrientDensity}
\alias{calcNutrientDensity}
\title{Calculate Feed Requirements}
\usage{
calcNutrientDensity()
}
\value{
MAgPIE-FEED data for NutrientDensity and corresonding weights as a
list of two MAgPIE objects
}
\description{
Provides MAgPIE-FEED data for Nutrient Density calculated in the regression
for feed (calcRegressionFEED). No changes to the content have been done.
Usually no weight needed as the data will be used in MAgPIE-FEED model which
is country based.
}
\examples{

\dontrun{ 
calcOutput("NutrientDensity")
}
}
\seealso{
\code{\link[madrat]{calcOutput}}, \code{\link{calcRegressionFEED}},
\code{\link[madrat]{readSource}}
}
\author{
Lavinia Baumstark, Isabelle Weindl
}
