% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcProdFoddr.R
\name{calcProdFoddr}
\alias{calcProdFoddr}
\title{Calculate FAO Fodder Crop Production aggregated to MAgPIE_FEED sectors}
\usage{
calcProdFoddr()
}
\value{
FAO fodder production data and corresponding weights as a list of
two MAgPIE objects
}
\description{
Provides the FAOSTAT Fodder production data (in mio ton) aggregated to three
MAgPIE_FEED fodder categories.
}
\examples{

\dontrun{ 
calcOutput("ProdFoddr")

}

}
\seealso{
\code{\link[madrat]{calcOutput}}, \code{\link{calcFAOFodder_aggrFEED}},
\code{\link[madrat]{readSource}}
}
\author{
Isabelle Weindl
}
