% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcProdSystRatio_2000.R
\name{calcProdSystRatio_2000}
\alias{calcProdSystRatio_2000}
\title{Calculate ProdSystRatio for 2000}
\usage{
calcProdSystRatio_2000()
}
\value{
MAgPIE[-FEED] data for ProdSystRatio and corresonding weights as a
list of two MAgPIE objects
}
\description{
Provides MAgPIE-FEED data for ProdSystRatio.Usually no weight needed as the
data will be used in MAgPIE-FEED model country based. Data aslo used as
input for MAgPIE, then an aggregation to regions is needed
}
\examples{

\dontrun{ 
calcOutput("ProdSystRatio_2000")

}

}
\seealso{
\code{\link[madrat]{calcOutput}}, \code{\link[madrat]{readSource}}
}
\author{
Lavinia Baumstark
}
