% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertWirseniusSubsystems.R
\name{convertWirseniusSubsystems}
\alias{convertWirseniusSubsystems}
\title{Convert Wirsenius Livestock Subsystems}
\usage{
convertWirseniusSubsystems(x, subtype)
}
\arguments{
\item{x}{the data to be converted?}

\item{subtype}{Type of subsystem information that should be read in. 
Available types are: "cattlemeat2milk", "meat2egg" and "feed_beefsys_ratio"}
}
\value{
magpie object of selected livestock subsystem data on country level
}
\description{
Convert data on livestock subsystems structure. Source: Wirsenius.
}
\examples{

\dontrun{ 
a <- readSource("WirseniusSubsystems","cattlemeat2milk")
a <- readSource("WirseniusSubsystems","meat2egg")
a <- readSource("WirseniusSubsystems","feed_beefsys")
}

}
\seealso{
\code{\link[madrat]{readSource}}
}
\author{
Isabelle Weindl
}
