% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrfeed-package.R
\docType{package}
\name{mrfeed-package}
\alias{mrfeed}
\alias{mrfeed-package}
\title{mrfeed: MadRat feed data package}
\description{
Provides feed related data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pik-piam/mrfeed}
  \item \doi{10.5281/zenodo.3822087}
  \item Report bugs at \url{https://github.com/pik-piam/mrfeed/issues}
}

}
\author{
\strong{Maintainer}: Isabelle Weindl \email{weindl@pik-potsdam.de}

Authors:
\itemize{
  \item Lavinia Baumstark \email{baumstark@pik-potsdam.de}
  \item Benjamin Leon Bodirsky \email{bodirsky@pik-potsdam.de}
  \item Nele Steinmetz
  \item Ulrich Kreidenweis
  \item Jan Philipp Dietrich \email{dietrich@pik-potsdam.de}
}

}
