% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readWirsenius.R
\name{readWirsenius}
\alias{readWirsenius}
\title{Read Wirsenius}
\usage{
readWirsenius()
}
\value{
magpie object of the Wirsenius data
}
\description{
Read-in an Wirsenius data .csv file as magclass object
}
\examples{

\dontrun{ a <- readSource(type="Wirsenius")
}

}
\seealso{
\code{\link[madrat]{readSource}}
}
\author{
Lavinia Baumstark
}
