% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLivestockProductivityWrapper.R
\name{calcLivestockProductivityWrapper}
\alias{calcLivestockProductivityWrapper}
\title{Calculate Feed Requirements}
\usage{
calcLivestockProductivityWrapper(future)
}
\arguments{
\item{future}{See \code{\link[mrcommons]{calcLivestockProductivity}}}
}
\value{
LivestockProductivity data
}
\description{
A wrapper for \code{\link[mrcommons]{calcLivestockProductivity}}, that sets the dimnames as required by the feedmodel
}
\examples{

\dontrun{ 
calcOutput("LivestockProductivityWrapper")

}
}
\author{
David Hötten
}
