% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullMAGPIE_FEED.R
\name{fullMAGPIE_FEED}
\alias{fullMAGPIE_FEED}
\title{fullMAGPIE_FEED}
\usage{
fullMAGPIE_FEED(rev = 0)
}
\arguments{
\item{rev}{data revision which should be used as input (positive numeric).}
}
\description{
Function that produces the complete country data set required for the
MAgPIE-FEED model (preprocessing).
}
\examples{
\dontrun{
fullMAGPIE_FEED()
}

}
\seealso{
\code{\link{readSource}}, \code{\link{getCalculations}}, \code{\link{calcOutput}}
}
\author{
Isabelle Weindl, Lavinia Baumstark
}
